+++ changelog.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Homepage structure complete
 - Portfolio structure complete
 - Markdown plugins
-- Platforms content
+- Platforms content and component
 
 ## [0.0.0] - 2019-01-14